refactoring legend tointo just one component
